[Refactor] 값 사용 구조 개선 - role값 enum으로 분리 (#17)
