+++ backend/src/main/java/org/example/backend/domain/company/controller/CompanyController.java
@@ -6,8 +6,8 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.example.backend.domain.company.model.dto.CompanyAuthTokenDto;
 import org.example.backend.domain.company.service.CompanyAuthTokenService;
-import org.example.backend.global.common.BaseResponse;
-import org.example.backend.global.common.BaseResponseStatus;
+import org.example.backend.global.common.constants.BaseResponse;
+import org.example.backend.global.common.constants.BaseResponseStatus;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
