+++ backend/src/main/java/org/example/backend/domain/user/controller/UserController.java
@@ -6,11 +6,10 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.example.backend.domain.user.model.dto.UserAuthTokenDto;
 import org.example.backend.domain.user.model.dto.UserDto;
-import org.example.backend.domain.user.model.entity.UserAuthToken;
 import org.example.backend.domain.user.service.UserAuthTokenService;
 import org.example.backend.domain.user.service.UserService;
-import org.example.backend.global.common.BaseResponse;
-import org.example.backend.global.common.BaseResponseStatus;
+import org.example.backend.global.common.constants.BaseResponse;
+import org.example.backend.global.common.constants.BaseResponseStatus;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
