+++ backend/src/main/java/org/example/backend/domain/user/service/UserService.java
@@ -7,9 +7,8 @@ import org.example.backend.domain.user.model.entity.User;
 import org.example.backend.domain.user.repository.DeliveryRepository;
 import org.example.backend.domain.user.repository.UserAuthTokenRepository;
 import org.example.backend.domain.user.repository.UserRepository;
-import org.example.backend.global.common.BaseResponseStatus;
+import org.example.backend.global.common.constants.BaseResponseStatus;
 import org.example.backend.global.exception.InvalidCustomException;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
