+++ backend/src/main/java/org/example/backend/global/common/constants/BaseResponse.java
@@ -1,4 +1,4 @@
-package org.example.backend.global.common;
+package org.example.backend.global.common.constants;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -6,7 +6,7 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
-import static org.example.backend.global.common.BaseResponseStatus.SUCCESS;
+import static org.example.backend.global.common.constants.BaseResponseStatus.SUCCESS;
 
 @Getter
 @AllArgsConstructor
