+++ backend/src/main/java/org/example/backend/global/exception/GlobalExceptionHandler.java
@@ -1,18 +1,12 @@
 package org.example.backend.global.exception;
 
 
-import jakarta.validation.ConstraintViolationException;
-import org.example.backend.global.common.BaseResponse;
+import org.example.backend.global.common.constants.BaseResponse;
 import org.springframework.core.annotation.Order;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.util.HashMap;
-import java.util.Map;
 
 @RestControllerAdvice
 @Order(value = 1)
